Increase CPU available to GCB job.
