+++ google-beta/data_source_google_kms_secret_ciphertext.go
@@ -13,8 +13,7 @@ import (
 
 func dataSourceGoogleKmsSecretCiphertext() *schema.Resource {
 	return &schema.Resource{
-		DeprecationMessage: "Use the google_kms_secret_ciphertext resource instead.",
-		Read:               dataSourceGoogleKmsSecretCiphertextRead,
+		Read: dataSourceGoogleKmsSecretCiphertextRead,
 		Schema: map[string]*schema.Schema{
 			"crypto_key": {
 				Type:     schema.TypeString,
