+++ website/docs/r/compute_backend_bucket_signed_url_key.html.markdown
@@ -94,6 +94,19 @@ This resource provides the following
 - `create` - Default is 4 minutes.
 - `delete` - Default is 4 minutes.
 
+## Import
+
+BackendBucketSignedUrlKey can be imported using any of these accepted formats:
+
+```
+$ terraform import google_compute_backend_bucket_signed_url_key.default projects/{{project}}/global/backendBuckets/{{backend_bucket}}/{{name}}
+$ terraform import google_compute_backend_bucket_signed_url_key.default {{project}}/{{backend_bucket}}/{{name}}
+$ terraform import google_compute_backend_bucket_signed_url_key.default {{backend_bucket}}/{{name}}
+```
+
+-> If you're importing a resource with beta features, make sure to include `-provider=google-beta`
+as an argument so that Terraform uses the correct provider to import your resource.
+
 ## User Project Overrides
 
 This resource supports [User Project Overrides](https://www.terraform.io/docs/providers/google/guides/provider_reference.html#user_project_override).
