+++ website/docs/r/compute_backend_service_signed_url_key.html.markdown
@@ -132,6 +132,19 @@ This resource provides the following
 - `create` - Default is 4 minutes.
 - `delete` - Default is 4 minutes.
 
+## Import
+
+BackendServiceSignedUrlKey can be imported using any of these accepted formats:
+
+```
+$ terraform import google_compute_backend_service_signed_url_key.default projects/{{project}}/global/backendServices/{{backend_service}}/{{name}}
+$ terraform import google_compute_backend_service_signed_url_key.default {{project}}/{{backend_service}}/{{name}}
+$ terraform import google_compute_backend_service_signed_url_key.default {{backend_service}}/{{name}}
+```
+
+-> If you're importing a resource with beta features, make sure to include `-provider=google-beta`
+as an argument so that Terraform uses the correct provider to import your resource.
+
 ## User Project Overrides
 
 This resource supports [User Project Overrides](https://www.terraform.io/docs/providers/google/guides/provider_reference.html#user_project_override).
