--- website/docs/r/kms_secret_ciphertext.html.markdown
@@ -1,126 +1,0 @@
----
-# ----------------------------------------------------------------------------
-#
-#     ***     AUTO GENERATED CODE    ***    AUTO GENERATED CODE     ***
-#
-# ----------------------------------------------------------------------------
-#
-#     This file is automatically generated by Magic Modules and manual
-#     changes will be clobbered when the file is regenerated.
-#
-#     Please read more about how to change this file in
-#     .github/CONTRIBUTING.md.
-#
-# ----------------------------------------------------------------------------
-subcategory: "Cloud KMS"
-layout: "google"
-page_title: "Google: google_kms_secret_ciphertext"
-sidebar_current: "docs-google-kms-secret-ciphertext"
-description: |-
-  Encrypts secret data with Google Cloud KMS and provides access to the ciphertext.
----
-
-# google\_kms\_secret\_ciphertext
-
-Encrypts secret data with Google Cloud KMS and provides access to the ciphertext.
-
-
-~> **NOTE**: Using this resource will allow you to conceal secret data within your
-resource definitions, but it does not take care of protecting that data in the
-logging output, plan output, or state output.  Please take care to secure your secret
-data outside of resource definitions.
-
-
-To get more information about SecretCiphertext, see:
-
-* [API documentation](https://cloud.google.com/kms/docs/reference/rest/v1/projects.locations.keyRings.cryptoKeys/encrypt)
-* How-to Guides
-    * [Encrypting and decrypting data with a symmetric key](https://cloud.google.com/kms/docs/encrypt-decrypt)
-
-## Example Usage - Kms Secret Ciphertext Basic
-
-
-```hcl
-resource "google_kms_key_ring" "keyring" {
-  name     = "keyring-example"
-  location = "global"
-}
-
-resource "google_kms_crypto_key" "cryptokey" {
-  name            = "crypto-key-example"
-  key_ring        = google_kms_key_ring.keyring.id
-  rotation_period = "100000s"
-
-  lifecycle {
-    prevent_destroy = true
-  }
-}
-
-resource "google_kms_secret_ciphertext" "my_password" {
-  crypto_key = google_kms_crypto_key.cryptokey.id
-  plaintext  = "my-secret-password"
-}
-
-resource "google_compute_instance" "instance" {
-  name         = "my-instance"
-  machine_type = "n1-standard-1"
-  zone         = "us-central1-a"
-
-  boot_disk {
-    initialize_params {
-      image = "debian-cloud/debian-9"
-    }
-  }
-
-  network_interface {
-    network = "default"
-
-    access_config {
-    }
-  }
-
-  metadata = {
-    password = google_kms_secret_ciphertext.my_password.ciphertext
-  }
-}
-```
-
-## Argument Reference
-
-The following arguments are supported:
-
-
-* `plaintext` -
-  (Required)
-  The plaintext to be encrypted.
-
-* `crypto_key` -
-  (Required)
-  The full name of the CryptoKey that will be used to encrypt the provided plaintext.
-  Format: `'projects/{{project}}/locations/{{location}}/keyRings/{{keyRing}}/cryptoKeys/{{cryptoKey}}'`
-
-
-- - -
-
-
-
-## Attributes Reference
-
-In addition to the arguments listed above, the following computed attributes are exported:
-
-
-* `ciphertext` -
-  Contains the result of encrypting the provided plaintext, encoded in base64.
-
-
-## Timeouts
-
-This resource provides the following
-[Timeouts](/docs/configuration/resources.html#timeouts) configuration options:
-
-- `create` - Default is 4 minutes.
-- `delete` - Default is 4 minutes.
-
-## User Project Overrides
-
-This resource supports [User Project Overrides](https://www.terraform.io/docs/providers/google/guides/provider_reference.html#user_project_override).
