+++ website/google.erb
@@ -1003,9 +1003,6 @@
       <li<%= sidebar_current("docs-google-kms-key-ring-iam") %>>
         <a href="/docs/providers/google/r/google_kms_key_ring_iam.html">google_kms_key_ring_iam_policy</a>
       </li>
-      <li<%= sidebar_current("docs-google-kms-secret-ciphertext") %>>
-        <a href="/docs/providers/google/r/kms_secret_ciphertext.html">google_kms_secret_ciphertext</a>
-      </li>
     </ul>
     </li>
